requirements.txt
affine==2.4.0
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
PySide2==5.15.2.1
pyxdg==0.28
rasterio==1.3.5.post1
shapely==2.0.1
shiboken2==5.15.2.1
show-in-file-manager==1.1.4
snuggs==1.4.7
